massive changes on the way that dates are stored on both implementations
now all date datatypes storre its contents in milliseconds on both
implementations (websql and indexeddb)
plus (this is very important) the milliseconds value is UTC'd, ie: when
you store the value 2015-10-10 it will converted to its UTC contrapart
based on your timezone offset, and this will be the final value that
msqta.orm will be store, and when you retrieve records based on a date,
this is will also converted to its UTC contrapart.
Finally in terms of user usage of the library, you dont have to change
anything, all these changes are internal and it dont affect the user
usage of the library.